# perf: delay compression in the bulk indexer