--- a/bulk_indexer.go
+++ b/bulk_indexer.go
@@ -19,13 +19,14 @@ package docappender
 
 import (
 	"bytes"
-	"compress/gzip"
 	"context"
 	"encoding/json"
 	"fmt"
 	"io"
 	"net/http"
+	"sync"
 
+	"github.com/klauspost/compress/gzip"
 	"go.elastic.co/fastjson"
 
 	"github.com/elastic/go-elasticsearch/v8"
@@ -61,11 +62,9 @@ type bulkIndexer struct {
 func newBulkIndexer(client *elasticsearch.Client, compressionLevel int) *bulkIndexer {
 	b := &bulkIndexer{client: client}
 	if compressionLevel != gzip.NoCompression {
-		b.gzipw, _ = gzip.NewWriterLevel(&b.buf, compressionLevel)
-		b.writer = b.gzipw
-	} else {
-		b.writer = &b.buf
+		b.gzipw, _ = gzip.NewWriterLevel(nil, compressionLevel)
 	}
+	b.writer = &b.buf
 	b.Reset()
 	return b
 }
@@ -75,7 +74,7 @@ func (b *bulkIndexer) Reset() {
 	b.itemsAdded, b.bytesFlushed = 0, 0
 	b.buf.Reset()
 	if b.gzipw != nil {
-		b.gzipw.Reset(&b.buf)
+		b.gzipw.Reset(nil)
 	}
 	b.resp = esutil.BulkIndexerResponse{Items: b.resp.Items[:0]}
 }
@@ -135,25 +134,39 @@ func (b *bulkIndexer) writeMeta(index, action, documentID string) {
 	b.jsonw.Reset()
 }
 
+var bufPool sync.Pool = sync.Pool{New: func() interface{} { return &bytes.Buffer{} }}
+
 // Flush executes a bulk request if there are any items buffered, and clears out the buffer.
 func (b *bulkIndexer) Flush(ctx context.Context) (esutil.BulkIndexerResponse, error) {
 	if b.itemsAdded == 0 {
 		return esutil.BulkIndexerResponse{}, nil
 	}
+
+	bbuf := &b.buf
+
 	if b.gzipw != nil {
+		buf := bufPool.Get().(*bytes.Buffer)
+		buf.Reset()
+		defer bufPool.Put(buf)
+
+		b.gzipw.Reset(buf)
+		b.gzipw.Write(b.buf.Bytes())
+
 		if err := b.gzipw.Close(); err != nil {
 			return esutil.BulkIndexerResponse{}, fmt.Errorf(
 				"failed closing the gzip writer: %w", err,
 			)
 		}
+
+		bbuf = buf
 	}
 
-	req := esapi.BulkRequest{Body: &b.buf, Header: make(http.Header)}
+	req := esapi.BulkRequest{Body: bbuf, Header: make(http.Header)}
 	if b.gzipw != nil {
 		req.Header.Set("Content-Encoding", "gzip")
 	}
 
-	bytesFlushed := b.buf.Len()
+	bytesFlushed := bbuf.Len()
 	res, err := req.Do(ctx, b.client)
 	if err != nil {
 		return esutil.BulkIndexerResponse{}, err