@kruskall 2022-12-13
I think the other PR broke when repository switched to public. I've reopened it here.

@marclop

> Micro benchmarks look good. I'm curious to see the performance improvement in practice. Did you run any macro benchmarks?

Yep, I run rally and as usual I'm getting inconsistent results. I repeated the test twice but that didn't help.

I cannot run the test more than twice because it is saturing my upload bandwith and it takes hours to finish.

Before:
- https://gist.github.com/kruskall/a6d775f88504c6fa34dedf0c69c03c66
- https://gist.github.com/kruskall/c4a9b23fea5fca6fef9fc70e46ba453e

After:
- https://gist.github.com/kruskall/794c553f678e18768431ce13ecdf9eed
- https://gist.github.com/kruskall/c337addad90aa6fa49eb5082a5cf67d1

>  I have some concerns on the behavior with the current FlushByte size (1MB) in the appender. Since we're not compressing the events when we add them to the cache, the cache will be filled up much and will also result in smaller bulk requests (less events).
> 
> We may need to increase the size of the cache along with this change to have an equal number of cached / flushed events per bulk request.

Yes, sorry, I mentioned that during the weekly and forgot to report it here. This is a change in behaviour, the "flush bytes" setting will refer to the uncompressed payload. We should increase the default setting and make sure to add a changelog stating that user might need to increase the flush bytes setting if they provided a custom one to account for the increased number of network requests